interface for secrets provider in a separate file
